implement formspree on signup page
There was the need of changing the css rule on style.css regarding
the form style. The css file was making reference to the HTML tag
which was blocking the tag name file change (from form to section).
The tag rename was required to not have nested forms as the action of
sending emails as a company that wants to hire and as a professional that
want to be hired are two complete different actions and needed a
form each.

The other minified css files were created when running gulp minification
and should not have any logic change.